+++ src/app.js
@@ -1,7 +1,7 @@
 
 const express = require('express')
 const app = express()
-const InMemoryWorkshop = require("./inMemoryWorkshop")
+const InMemoryWorkshop = require("./services/workshops")
 const path = require("path")
 const ejs = require('ejs')
 var bodyParser = require('body-parser')
@@ -10,14 +10,14 @@ app.use(bodyParser.urlencoded({ extended: false }))
 
 // set the view engine to ejs
 app.set('view engine', 'ejs');
-app.set('views', path.join(__dirname, '..', '/ejs'));
-app.use(express.static(path.join(__dirname , '..', 'css')));
+app.set('views', path.join(__dirname, '.', 'views/pages'));
+app.use(express.static(path.join(__dirname , '.', 'views/style')));
 
 
 app.get('/', function (req, res) {
     InMemoryWorkshop.getWorkshopList()
     .then(workshops => {
-        res.render("index", {
+        res.render("app", {
             workshops: workshops
         })
     })
@@ -34,7 +34,7 @@ app.post('/workshop', function (req, res) {
     InMemoryWorkshop.addWorkshop(name, description).then(() => {
         InMemoryWorkshop.getWorkshopList()
         .then(workshops => {
-            res.render("index", {
+            res.render("app", {
                 workshops: workshops
             })
         })
@@ -46,7 +46,7 @@ app.get('/workshop/:name', function (req, res) {
     const workshopName = req.params.name
     InMemoryWorkshop.getWorkshopByName(workshopName)
     .then(workshop => {
-        res.render('ejs/workshop', workshop)
+        res.render('workshop', workshop)
     })
     .catch(e =>ejs.send(e.message))
 })
